--- Sources/Fluent/Service/Databases.swift
@@ -1,5 +1,0 @@
-/// Represents the databases currently configured for Fluent.
-public struct Databases {
-    /// Internal storage: [DatabaseIdentifier: Database]
-    internal let storage: [String: Any]
-}
